refactor(types): select
