+++ src/lib/builders/select/types.ts
@@ -1,0 +1,22 @@
+import type { FloatingConfig } from '@melt-ui/svelte/internal/actions';
+import type { createSelect } from './create';
+
+export type CreateSelectArgs = {
+	positioning?: FloatingConfig;
+	arrowSize?: number;
+	required?: boolean;
+	disabled?: boolean;
+	value?: unknown;
+	label?: string;
+	name?: string;
+	preventScroll?: boolean;
+	loop?: boolean;
+};
+
+export type OptionArgs = {
+	value: unknown;
+	label?: string;
+	disabled?: boolean;
+};
+
+export type CreateSelectReturn = ReturnType<typeof createSelect>;
